Update to match wordpress.org version
Updated tab ids to match the group ids that are in the database.
Using the same version of the jQuery UI theme as the jQuery UI installed
in the current version of WordPress on the server.
Added example tabs in the admin section.
Registering scripts based on http or https protocol to avoid https
errors.